Test audit policy effect
To demostrate how audit effect can be tested.

Checking effects different from `deny` it's slower since the
policy evaluation results take quite a long time
(in the order of minutes) to be available.

In order to check that `audit` effects are correctly applied, the
policy evaluation is triggered and then the compliance state of the
resource should be `NonCompliant`.

The code calls the Azure rest API directly instead of using the
[Go SDK](https://github.com/Azure/azure-sdk-for-go). There are two reasons:

- No extensive use of the Azure API, only two endpoints are called;
- opportunity to experiment with some Go features.